requirements.txt
-e git://github.com/PaulUithol/backbone-tastypie.git@f253883cc908a643f3b017f5cd848f3180d75567#egg=Backbone_tastypie-dev
Django==1.4.2
argparse==1.2.1
django-bootstrap-toolkit==2.6.16
django-tastypie==0.9.11
django-verbatim==0.1
mimeparse==0.1.3
python-dateutil==1.5
requests==0.14.2
wadofstuff-django-serializers==1.1.0
wsgiref==0.1.2
